释放部分过度封装的方法
