+++ package.json
@@ -21,7 +21,7 @@
         "components"
     ],
     "directories": {
-        "jtool": "^1.2.26"
+        "jtool": "^1.2.28"
     },
     "devDependencies": {
         "babel-core": "^6.25.0",
@@ -77,5 +77,7 @@
     "peerDependencies": {
         "library": "^1.3.5"
     },
-    "dependencies": {}
+    "dependencies": {
+        "jtool": "^1.2.28"
+    }
 }
