+++ src/common/Store.js
@@ -2,7 +2,7 @@
  * Created by baukh on 17/10/24.
  * 实例化数据的存储对象
  */
-import { GM_VERSION } from '../common/constants';
+import { GM_VERSION } from './constants';
 const Store = {
 	// 版本号
 	version: GM_VERSION,
