+++ src/common/index.js
@@ -1,0 +1,9 @@
+/*
+ * 公共方法
+ * */
+import jTool from './jTool';
+import { parseTpl } from './parse';
+import base from './base';
+import cache from './cache';
+
+export { jTool, base, cache, parseTpl };
