+++ src/common/jTool.js
@@ -1,0 +1,6 @@
+import 'jtool';
+let jTool = window.jTool;
+
+window.jTool === window.$ && delete window.$;
+delete window.jTool;
+export default jTool;
