+++ src/module/menu/index.js
@@ -1,8 +1,7 @@
 /*
  * GridManager: 右键菜单
  * */
-import { jTool, base, parseTpl } from '../base';
-import cache from '../cache';
+import { jTool, base, cache, parseTpl } from '../../common';
 import i18n from '../i18n';
 import exportFile from '../exportFile';
 import ajaxPage from '../ajaxPage';
